Emacs: org-super-agenda
Mainly configured org-super-agenda for a nicer agenda-view. Also changed
the fonts for Nanorocket and switched to the =Nord= theme.
Todo-states where also changed to include =FILE= for finished tasks that
still need to be documented.
Finally added a proper org-capture template for contacts.
Changed auth-source to password-store.
Added pass in =doom/init.el= and added keybindings for ~pass~ and ~ivy-pass~.